Make the github repository structure more clear
