+++ docker-compose.yml
@@ -25,6 +25,7 @@ services:
     volumes:
       - pg_data:/var/lib/postgresql/data
       - ./db/init.sql:/docker-entrypoint-initdb.d/init.sql
+      - ./db/data:/data
     healthcheck:
       test: ["CMD-SHELL", "pg_isready -U dwv -d dwv"]
       interval: 10s
@@ -70,7 +71,5 @@ services:
     command: ["python", "scheduler.py"]
 
 volumes:
-  mongo_data:
-  ch_data:
   pg_data:
   pgadmin_data:
